Implement API-NBA search error handling
Added check for a size '0' player array; as the api does not throw an
error when a player's last name is not found in the database.

As a result no data is delivered to the user and no message is given
to explain failure.

Now if the player array's lenght is '0' an error will be thrown
promting the user to try again.