Refactor social searcher api pull
API requires a header to pull json; but will not accept the API key
within the header. At least the documents do not make it clear in how
to so.

Will reach out to the support team to see if auth can be passed thorugh
the header for more secure access.