Update Player and Team data pull
Previous commits have not saved due to an error in gitBash.

Pulling player data and displaying general data to the user via the
DOM. This does not come without issues; the api's return data when
pulling a player does not include the team name.

The return data only includes the teamId; which is difficult to work
with as I do not know of a way to run an async function within another
async function.

Team data displys properly; a future decision may involve limiting the
team search to the NBA only. Currently the data pull will include
international teams.