Remove break from non NBA leagues in player search
The API-NBA API includes non-US leagues and players; this causes a
break when searching and pulling position. Given this the loop will
skip over any player not containing the 'standard' league key.

This needs to be replicated in the team search to remove the population
of All Star and International teams.