Refactor get functions and store search results
Added a transition to take place when clicking a search result to bring
additional data onto the screen that dives deeper into that search
result. Added a hidden element to store the id of the items displayed
so that data can be pulled when needing to populate additional details.

Search results are stored in a global array to access and avoid
multiple pulls from the server for the same data.

Refactored the conference name function to be standalone for easy
access site wide.