Refactor NBA player and team functions
Considering the API-NBA api does not include team data outsie of the
team ID an additional fetch is required to obtain all the data
necessary. To avoid multiple data sets I refactored with the spread
opporator to ensure all data is under the same object when search
results for players are displayed.

This same approach will be used to truncate the multiple fetch calls
needed to pull different social media platform posts for a give NBA
item.