Updated app to display player data after selection
Added functionality to read player and display the player's profile
data based on the users player selection following their search.

This functionality also allows for continuous search without requiring
the user to go back a page.

Additional  functions are needed to display the height, weight and date
of birth in the US standard.