Add players from given team to the DOM
When a player has been selected by the user in addition to the videos,
and news; players from the same team will be displayed to provide an
ideal user experience.

Additionally the social searcher API has an error that is preventing
data from returning properly. A 200 status is received when making a
request. This will be addressed directly with social searcher API.