Add team search functionality and population
Added team display functionality to display general team infomraiton,
along with news, social media, videos and players on the team with full
integration to click on all assets and end up in the right location.

Both the population of players and social media have randomness so the
look and feel is different every time.

Addditional data will be added in future version such as game data and
potentially some stats updates for players. This all depends on the
flexibility of the API.