Update conference page to display divisions
To provide a deeper user expereience an option to display divisions as
a branch off from a given conference is being developed. Currently
functionality is not there to actually display the division; just
placeholders for the clickable buttons to access those pages within
each conference.

Once complete this will compelte the development of the MVP for general
user testing prior to styling the application.