Complete division display functionatlity
Added division display to the conference view allowing the user to dive
deeper into the conference and see division specific news and content.

This completes the baseline functionality needed for the MVP; there are
obvious improvements that could take place but this version will allow
for valuable user feedback to improve the application as it grows in the
future.