Fix error and add loading notice
Addressed an issue with truncating the author name when needed; this was
causing premature loading of articles and resulting in 5 articles not
being displayed.

Added loading text to the social media and team assets sections to
inform the user content is still loading. This should prevent blank
space and giving the impression that there is an error.