Revamp layout
Changed #connectedItems to a combination of grid and flex to have a
better representation of the players and other connected items to the
user's current search selection.

This adds a better layout and prevents the creation of additional white
space when it is not needed given the amount of players or teams being
populated in that section.

Altered the first row of each section to better support the title
structure and prevent additional white space.

Side note: attempted to add clientside form validation to the search
field. For some reason the form will not submit when the pattern
requested has been forfilled. This issue will be addressed; but will be
removed in a future commit in order to move forward with other
development features.