Restyle userSelection section
Altered teh grid to show a player, team or conference to display the
video section in a greater manner. Providing the theater video display
feel.

This adjustment also included the integration of a hover effect on the
subvideos for the user to navigate to.

future inprovement will include a video switcher and embed all the video
player directly in the page without having to link to YouTube directly.