Complete loading animation
Added facebook like loading animation to avoid obvious loading content;
this feature exists when searching and while a player, team or
conference is loading.

This will show in the social and connected items sections considering
the amount of time needed to make all fetch calls to complete the data
in those areas.

Next is to complete mobile responsiveness.